updates doc strings, documentation, and moves function locations
