+++ docs/source/reference/index.md
@@ -1,6 +1,20 @@
 # API Reference
 
+## Model Classes
 ```{eval-rst}
+
+.. autosummary::
+   :toctree: generated
+   :nosignatures:
+   :template: my_class.rst
+
+    osier.DispatchModel
+    osier.CapacityExpansion
+```
+
+## Technology Classes
+```{eval-rst}
+
 .. autosummary::
    :toctree: generated
    :nosignatures:
@@ -10,10 +24,28 @@
     osier.RampingTechnology
     osier.ThermalTechnology
     osier.StorageTechnology
-    osier.DispatchModel
-    osier.CapacityExpansion
-    osier.utils.synchronize_units
-    osier.equations.per_unit_energy
-    osier.equations.per_unit_capacity
-    osier.equations.volatility
+
+```
+
+## Objective Functions
+```{eval-rst}
+.. autosummary::
+   :toctree: generated
+   :nosignatures:
+   :template: custom_module_template.rst
+
+    osier.equations
+
+```
+
+
+## General Functions
+```{eval-rst}
+.. autosummary::
+   :toctree: generated
+   :nosignatures:
+   :template: custom_module_template.rst
+
+    osier.utils
+
 ```
