+++ osier/equations.py
@@ -1,89 +1,8 @@
-from abc import ABC, abstractmethod
+from osier.utils import (get_tech_names, 
+                         get_dispatchable_techs, 
+                         get_nondispatchable_techs, get_dispatchable_names)
 from pyentrp.entropy import weighted_permutation_entropy
 import numpy as np
-
-
-class OsierEquation(ABC):
-
-    def __init__(self) -> None:
-        pass
-
-    @abstractmethod
-    def _do(self, technology_list, X):
-        pass
-
-
-def get_tech_names(technology_list):
-    """
-    Parameters
-    ----------
-    technology_list : list of :class:`osier.Technology` objects
-        The list of technologies.
-
-    Returns
-    -------
-    tech_names : list of str
-        The list of technology names.
-    """
-
-    tech_names = [t.technology_name for t in technology_list]
-
-    return tech_names
-
-
-def get_dispatchable_techs(technology_list):
-    """
-    Parameters
-    ----------
-    technology_list : list of :class:`osier.Technology` objects
-        The list of technologies.
-
-    Returns
-    -------
-    tech_names : list of :class:`osier.Technology`
-        The list of dispatchable technologies.
-    """
-
-    dispatchable_techs = [t for t in technology_list if t.dispatchable]
-
-    return dispatchable_techs
-
-
-def get_nondispatchable_techs(technology_list):
-    """
-    Parameters
-    ----------
-    technology_list : list of :class:`osier.Technology` objects
-        The list of technologies.
-
-    Returns
-    -------
-    non_dispatchable_techs : list of :class:`osier.Technology`
-        The list of non dispatchable technologies.
-    """
-
-    non_dispatchable_techs = [t for t in technology_list if not t.dispatchable]
-
-    return non_dispatchable_techs
-
-
-def get_dispatchable_names(technology_list):
-    """
-    Parameters
-    ----------
-    technology_list : list of :class:`osier.Technology` objects
-        The list of technologies.
-
-    Returns
-    -------
-    dispatchable_names : list of str
-        The list of dispatchable technology names.
-    """
-
-    dispatchable_names = [
-        t.technology_name for t in technology_list if t.dispatchable]
-
-    return dispatchable_names
 
 
 def per_unit_capacity(technology_list,
@@ -93,6 +12,19 @@ def per_unit_capacity(technology_list,
     This function calculates a general objective for a given
     set of technologies and their corresponding dispatch on a
     per-unit-capacity basis.
+
+    The general objective function is
+
+    .. math::
+        \\mathcal{K} = \\sum_g^G \\textbf{CAP}_g \\kappa_g,
+
+    .. math::
+        \\textbf{CAP}_g = \\text{the capacity of the g-th technology}
+        \\quad \\left[MW\\right].
+        
+    .. math::
+        \\kappa_g = \\text{the power density of the g-th technology} 
+        \\quad \\left[\\frac{-}{MW}\\right].
 
     .. warning::
         User-defined attributes are not validated by :class:`osier`.
@@ -137,6 +69,20 @@ def per_unit_energy(technology_list, attribute, solved_dispatch_model):
     set of technologies and their corresponding dispatch on a
     per-unit-energy basis.
 
+    The general objective function is 
+     
+    .. math::
+        \\mathcal{X} = \\sum_g^G \\xi_g \\sum_t^T x_{g,t},
+    
+    .. math::
+        x_{g,t} = \\text{the energy produced by the g-th technology
+         at time t}\\quad \\left[MWh\\right]
+
+    .. math::
+        \\xi_g = \\text{the energy density of the g-th technology}\\quad
+        \\left[\\frac{-}{MWh}\\right].
+
+        
     .. warning::
         User-defined attributes are not validated by :class:`osier`.
         Verify the units are accurate and uniform across all
